+++ glib/glib_test.go
@@ -1,8 +1,8 @@
 package glib_test
 
 import (
-	"github.com/conformal/gotk3/glib"
-	"github.com/conformal/gotk3/gtk"
+	"github.com/envoker/gotk3/glib"
+	"github.com/envoker/gotk3/gtk"
 	"runtime"
 	"testing"
 )
