+++ gtk/examples/simple/simple.go
@@ -19,7 +19,7 @@
 package main
 
 import (
-	"github.com/conformal/gotk3/gtk"
+	"github.com/envoker/gotk3/gtk"
 	"log"
 )
