+++ gtk/gtk_3_10-12.go
@@ -29,8 +29,8 @@ package gtk
 // #include "gtk_3_10-12.go.h"
 import "C"
 import (
-	"github.com/conformal/gotk3/gdk"
-	"github.com/conformal/gotk3/glib"
+	"github.com/envoker/gotk3/gdk"
+	"github.com/envoker/gotk3/glib"
 	"runtime"
 	"unsafe"
 )
