+++ gtk/gtk_3_6-10.go
@@ -27,7 +27,7 @@ package gtk
 // #include <gtk/gtk.h>
 import "C"
 import (
-	"github.com/conformal/gotk3/glib"
+	"github.com/envoker/gotk3/glib"
 	"runtime"
 	"unsafe"
 )
