+++ gtk/gtk_3_6-8.go
@@ -28,7 +28,7 @@ package gtk
 // #include <gtk/gtk.h>
 import "C"
 import (
-	"github.com/conformal/gotk3/glib"
+	"github.com/envoker/gotk3/glib"
 	"runtime"
 	"unsafe"
 )
