+++ pango/pango.go
@@ -21,7 +21,7 @@ package pango
 // #include <pango/pango.h>
 import "C"
 import (
-	"github.com/conformal/gotk3/glib"
+	"github.com/envoker/gotk3/glib"
 	"unsafe"
 )
